Python:Rename 'brave_search.py' to 'brave.py' and update the reference path in other files as well
