+++ python/samples/concepts/search/brave_text_search_as_plugin.py
@@ -5,7 +5,7 @@ from collections.abc import Awaitable, Callable
 from semantic_kernel import Kernel
 from semantic_kernel.connectors.ai import FunctionChoiceBehavior
 from semantic_kernel.connectors.ai.open_ai import OpenAIChatCompletion, OpenAIChatPromptExecutionSettings
-from semantic_kernel.connectors.search.brave_search import BraveSearch
+from semantic_kernel.connectors.search.brave import BraveSearch
 from semantic_kernel.contents import ChatHistory
 from semantic_kernel.filters import FilterTypes, FunctionInvocationContext
 from semantic_kernel.functions import KernelArguments, KernelParameterMetadata, KernelPlugin
